feat: drop Validator Interface and methods
In the past, Validator interface was added to implement
client side validation of Kong entities.

Overtime, Kong entities have undergone a lot of change and become
much more complicated.
Replicating the schema check library in Go code
will complicate things, introduce more bugs and won't buy us much.
We have already been bitten by this solution:
https://github.com/hbagdi/deck/issues/44

Hence, this is being dropped from the library.

In future, if this is needed, a better solution would be to either
re-use schema library of Kong here or at least re-use the schema of Kong
entities, that is exposed on the Admin API as of Kong 1.2.